+++ examples/resources/forgejo_repository_actions_secret/resource.tf
@@ -1,0 +1,6 @@
+resource "forgejo_repository_actions_secret" {
+  data       = "secret"
+  name       = "test"
+  owner      = "adyxax"
+  repository = "example"
+}