+++ internal/provider/provider.go
@@ -75,7 +75,9 @@ func (p *Provider) Configure(ctx context.Context, req provider.ConfigureRequest,
 }
 
 func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
-	return []func() resource.Resource{}
+	return []func() resource.Resource{
+		NewRepositoryActionsSecretResource,
+	}
 }
 
 func (p *Provider) DataSources(ctx context.Context) []func() datasource.DataSource {